Fixed compilation errors
* Moved shared (duplicated) utility functions to a separate file

* Fixed syntax errors and misnamed variables
