+++ Code/PocketMage_V3/src/CALENDAR.cpp
@@ -215,17 +215,6 @@ int stringToPositiveInt(String input) {
   return input.toInt();
 }
 
-int daysInMonth(int year, int month) {
-  if (month == 2) {
-    // Leap year
-    return (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)) ? 29 : 28;
-  } else if (month == 4 || month == 6 || month == 9 || month == 11) {
-    return 30;
-  } else {
-    return 31;
-  }
-}
-
 void commandSelectMonth(String command) {
   command.toLowerCase();
